Fixed positioning bug + cleanup and formatting
