--- Assets/ToolbarExtender/Example/Scripts/Editor/SceneHelper.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 5becded864e6414b983cc9378dae8155
-timeCreated: 1534172438
