--- Assets/ToolbarExtender/Scripts/Editor/ExtendedToolbarWindow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a5542090de7e848edbaad9dc3b8fa2e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
