+++ Assets/ToolbarExtender/Scripts/Editor/ToolbarCallback.cs
@@ -1,60 +1,61 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using UnityEngine;
 using UnityEditor;
 using System.Reflection;
 using UnityEngine.Experimental.UIElements;
 
-public static class ToolbarCallback
+namespace UnityToolbarExtender
 {
-    static Type m_toolbarType = typeof(Editor).Assembly.GetType("UnityEditor.Toolbar");
-    static Type m_guiViewType = typeof(Editor).Assembly.GetType("UnityEditor.GUIView");
-    static PropertyInfo m_viewVisualTree = m_guiViewType.GetProperty("visualTree", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
-    static FieldInfo m_imguiContainerOnGui = typeof(IMGUIContainer).GetField("m_OnGUIHandler", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
+	public static class ToolbarCallback
+	{
+		static Type m_toolbarType = typeof(Editor).Assembly.GetType("UnityEditor.Toolbar");
+		static Type m_guiViewType = typeof(Editor).Assembly.GetType("UnityEditor.GUIView");
+		static PropertyInfo m_viewVisualTree = m_guiViewType.GetProperty("visualTree",
+			BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
+		static FieldInfo m_imguiContainerOnGui = typeof(IMGUIContainer).GetField("m_OnGUIHandler",
+			BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
+		static ScriptableObject m_currentToolbar;
 
-    static ScriptableObject m_currentToolbar;
+		/// <summary>
+		/// Callback for toolbar OnGUI method.
+		/// </summary>
+		public static Action OnToolbarGUI;
 
-    /// <summary>
-    /// Callback for toolbar OnGUI method.
-    /// </summary>
-    public static Action OnToolbarGUI;
+		static ToolbarCallback()
+		{
+			EditorApplication.update -= OnUpdate;
+			EditorApplication.update += OnUpdate;
+		}
 
-    static ToolbarCallback()
-    {
-        EditorApplication.update -= OnUpdate;
-        EditorApplication.update += OnUpdate;
-    }
+		static void OnUpdate()
+		{
+			// Relying on the fact that toolbar is ScriptableObject and gets deleted when layout changes
+			if (m_currentToolbar == null)
+			{
+				// Find toolbar
+				var toolbars = Resources.FindObjectsOfTypeAll(m_toolbarType);
+				m_currentToolbar = toolbars.Length > 0 ? (ScriptableObject) toolbars[0] : null;
+				if (m_currentToolbar != null)
+				{
+					// Get it's visual tree
+					var visualTree = (VisualElement) m_viewVisualTree.GetValue(m_currentToolbar, null);
 
-    static void OnUpdate()
-    {
-        // Relying on the fact that toolbar is ScriptableObject and gets deleted when layout changes
-        if (m_currentToolbar == null)
-        {
-            // Find toolbar
-            var toolbars = Resources.FindObjectsOfTypeAll(m_toolbarType);
-            m_currentToolbar = toolbars.Length > 0 ? (ScriptableObject)toolbars[0] : null;
-            if (m_currentToolbar != null)
-            {
-                // Get it's visual tree
-                var visualTree = (VisualElement)m_viewVisualTree.GetValue(m_currentToolbar, null);
+					// Get first child which 'happens' to be toolbar IMGUIContainer
+					var container = (IMGUIContainer) visualTree[0];
 
-                // Get first child which 'happens' to be toolbar IMGUIContainer
-                var container = (IMGUIContainer)visualTree[0];
+					// (Re)attach handler
+					var handler = (Action) m_imguiContainerOnGui.GetValue(container);
+					handler -= OnGUI;
+					handler += OnGUI;
+					m_imguiContainerOnGui.SetValue(container, handler);
+				}
+			}
+		}
 
-                // (Re)attach handler
-                var handler = (Action)m_imguiContainerOnGui.GetValue(container);
-                handler -= OnGUI;
-                handler += OnGUI;
-                m_imguiContainerOnGui.SetValue(container, handler);
-            }
-        }
-    }
-
-    static void OnGUI()
-    {
-        var handler = OnToolbarGUI;
-        if (handler != null) handler();
-    }
+		static void OnGUI()
+		{
+			var handler = OnToolbarGUI;
+			if (handler != null) handler();
+		}
+	}
 }
